requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
decorator==5.1.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
validators==0.20.0
